Apply MCP 1003 and move `diagnostics.rs` into its own module and move `error_helper.rs` into the `diagnostics` folder in `rustc_resolve`
