+++ compiler/rustc_resolve/src/build_reduced_graph.rs
@@ -31,7 +31,7 @@ use tracing::debug;
 
 use crate::Namespace::{MacroNS, TypeNS, ValueNS};
 use crate::def_collector::DefCollector;
-use crate::error_helper::{OnUnknownData, StructCtor};
+use crate::diagnostics::impls::{OnUnknownData, StructCtor};
 use crate::imports::{ImportData, ImportKind};
 use crate::macros::{MacroRulesDecl, MacroRulesScope, MacroRulesScopeRef};
 use crate::ref_mut::CmCell;
